requirements.txt
Jinja2==2.10.3
SQLAlchemy==1.3.11
Werkzeug==0.16.0
pytz==2019.3
sqlalchemy-migrate==0.13.0
requests==2.22.0
Flask==1.1.1
Telethon==1.10.8
mysql-connector-python==8.0.18
gspread==3.1.0
oauth2client==4.1.3
pyyaml==5.2